got checkout list to work and persist.
will need to fix up some data types I made so that it is easier to
manage code, as well as the delete property for students, and ability to
add and edit classes in the payment tables. will also need to make sure
the checkmarks also change the attendances for students in payment
tables, and that checking students from other days will affect the same
list.